Add support for exfat file systems.
Decided to go for the solution presented by
the github.com/arter97/exfat-linux project. It
is a backport of the exfat module for our kernel
version. The exfat repository was added as a git submodule to
the kernel tree under the path fs/exfat.

Added necessary configuration variables to fs/Kconfig and fs/Makefile.

Change-Id: I3fb43eeeaa210d215846a1ebcc10c6a8a075dc93